log updates

+++ db/accounting-ali/addtoStockPrimList.php
@@ -25,15 +25,15 @@ if ($result = mysqli_query($con, $sql)) {
         $invoiceData[$cr2]['customerName'] = $row['customerName'];
         $invoiceData[$cr2]['customerId'] = $row['customerId'];
         $invoiceData[$cr2]['transactionType'] = $row['transactionType'];
-        $invoiceData[$cr2]['price'] =  (int)$row['price'];
+        $invoiceData[$cr2]['price'] =  (float)$row['price'];
         $invoiceData[$cr2]['Qty'] = (int)$row['Qty'];
         $invoiceData[$cr2]['invNumber'] = (int)$row['invNumber'];
         $invoiceData[$cr2]['date_time'] = $row['date_time'];
         $invoiceData[$cr2]['notes'] = $row['notes'];
-        $invoiceData[$cr2]['invoiceTotal'] = (int)$row['invoiceTotal'];     
+        $invoiceData[$cr2]['invoiceTotal'] = (float)$row['invoiceTotal'];     
         $cr2++;
     }
-    
+
     echo json_encode($invoiceData);
 } else {
     http_response_code(404);
